Modify generated job runner to use cli.entry_point
